pyproject.toml
[project]
name = "django-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]
